+++ packages/api-schema/src/index.ts
@@ -20,7 +20,7 @@ export type AuthGetMeApiSchema = CreateApiSchema<{
     name: string,
     picture: string,
   },
-}>; 
+}>;
 
 export type AuthGetLogoutApiSchema = CreateApiSchema<{
   res: {
@@ -35,3 +35,48 @@ export type FileGetApiSchema = CreateApiSchema<{
     originalName: string,
   },
 }>;
+
+// For Registration form data types
+
+export type StudentFormData = {
+  prefixTH: string;
+  nameTH: string;
+  middleNameTH: string;
+  surnameTH: string;
+  prefixEN: string;
+  nameEN: string;
+  middleNameEN: string;
+  surnameEN: string;
+  nicknameTH: string;
+  grade: string;
+  coolQuote: string;
+  email: string;
+  phone: string;
+  line: string;
+  preferFood: string;
+  allergyFood: string;
+  allergyDrug: string;
+  medicalProblem: string;
+};
+
+export type TeamFormData = {
+  teamName: string;
+  school: string;
+  amount: number;
+  prefixTH: string;
+  nameTH: string;
+  middleNameTH: string;
+  surnameTH: string;
+  prefixEN: string;
+  nameEN: string;
+  middleNameEN: string;
+  surnameEN: string;
+  email: string;
+  phone: string;
+  line: string;
+};
+
+export type RegistrationFormData = {
+  students: StudentFormData[],
+  team: TeamFormData,
+};
